Split out the generation of new rage events into separate file
